[Refactor] feature.cpp/h に定義されていた関数群をgrid.cpp/h に移設してファイルは削除した
