+++ src/floor/wild.cpp
@@ -16,7 +16,6 @@
 #include "floor/cave.h"
 #include "game-option/birth-options.h"
 #include "game-option/map-screen-options.h"
-#include "grid/feature.h"
 #include "grid/grid.h"
 #include "info-reader/fixed-map-parser.h"
 #include "info-reader/parse-error-types.h"
