+++ src/grid/grid.h
@@ -66,6 +66,9 @@ class PlayerType;
 class MonraceDefinition;
 enum class GridCountKind;
 enum class TerrainCharacteristics;
+enum class TerrainTag;
+void cave_set_feat(PlayerType *player_ptr, const Pos2D &pos, TerrainTag tag);
+void cave_set_feat(PlayerType *player_ptr, const Pos2D &pos, short terrain_id);
 bool new_player_spot(PlayerType *player_ptr);
 bool player_can_enter(PlayerType *player_ptr, FEAT_IDX feature, BIT_FLAGS16 mode);
 void update_local_illumination(PlayerType *player_ptr, POSITION y, POSITION x);
