+++ src/room/rooms-city.cpp
@@ -1,20 +1,15 @@
 #include "room/rooms-city.h"
-#include "floor/floor-generator.h"
 #include "game-option/cheat-types.h"
-#include "grid/feature.h"
 #include "grid/grid.h"
 #include "room/space-finder.h"
-#include "store/store-util.h"
 #include "store/store.h"
 #include "system/angband-exceptions.h"
 #include "system/floor/floor-info.h"
-#include "system/floor/town-info.h"
 #include "system/floor/town-list.h"
 #include "system/grid-type-definition.h"
 #include "system/player-type-definition.h"
 #include "system/terrain/terrain-definition.h"
 #include "system/terrain/terrain-list.h"
-#include "util/bit-flags-calculator.h"
 #include "wizard/wizard-messages.h"
 #include <algorithm>
