+++ src/wizard/wizard-special-process.cpp
@@ -12,9 +12,6 @@
 #include "wizard/wizard-special-process.h"
 #include "artifact/fixed-art-generator.h"
 #include "artifact/fixed-art-types.h"
-#include "birth/inventory-initializer.h"
-#include "cmd-io/cmd-dump.h"
-#include "cmd-io/cmd-help.h"
 #include "cmd-io/cmd-save.h"
 #include "cmd-visual/cmd-draw.h"
 #include "core/asking-player.h"
@@ -30,27 +27,15 @@
 #include "game-option/option-types-table.h"
 #include "game-option/play-record-options.h"
 #include "game-option/special-options.h"
-#include "grid/feature.h"
 #include "grid/grid.h"
-#include "info-reader/fixed-map-parser.h"
-#include "inventory/inventory-object.h"
-#include "inventory/inventory-slot-types.h"
 #include "io/files-util.h"
 #include "io/input-key-requester.h"
 #include "io/write-diary.h"
-#include "market/arena.h"
 #include "monster-floor/monster-remover.h"
-#include "monster-floor/monster-summon.h"
 #include "monster/monster-describer.h"
 #include "monster/monster-description-types.h"
-#include "monster/monster-info.h"
-#include "monster/monster-status.h"
-#include "monster/smart-learn-types.h"
-#include "mutation/mutation-investor-remover.h"
 #include "object-enchant/item-apply-magic.h"
 #include "object-enchant/item-magic-applier.h"
-#include "object-enchant/trc-types.h"
-#include "object-enchant/trg-types.h"
 #include "perception/object-perception.h"
 #include "player-base/player-class.h"
 #include "player-base/player-race.h"
@@ -60,27 +45,16 @@
 #include "player-info/self-info.h"
 #include "player-status/player-energy.h"
 #include "player/digestion-processor.h"
-#include "player/patron.h"
 #include "player/player-realm.h"
 #include "player/player-skill.h"
 #include "player/player-spell-status.h"
 #include "player/player-status-table.h"
 #include "player/player-status.h"
 #include "player/race-info-table.h"
-#include "spell-kind/spells-detection.h"
-#include "spell-kind/spells-sight.h"
-#include "spell-kind/spells-teleport.h"
 #include "spell-kind/spells-world.h"
-#include "spell/spells-object.h"
 #include "spell/spells-status.h"
-#include "spell/spells-summon.h"
 #include "status/bad-status-setter.h"
-#include "status/experience.h"
-#include "system/angband-system.h"
-#include "system/angband-version.h"
 #include "system/artifact-type-definition.h"
-#include "system/baseitem/baseitem-definition.h"
-#include "system/baseitem/baseitem-list.h"
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
@@ -88,26 +62,18 @@
 #include "system/grid-type-definition.h"
 #include "system/item-entity.h"
 #include "system/monster-entity.h"
-#include "system/player-type-definition.h"
 #include "system/redrawing-flags-updater.h"
 #include "system/terrain/terrain-definition.h"
 #include "system/terrain/terrain-list.h"
 #include "target/grid-selector.h"
-#include "term/screen-processor.h"
-#include "term/z-form.h"
 #include "util/angband-files.h"
 #include "util/bit-flags-calculator.h"
 #include "util/candidate-selector.h"
-#include "util/enum-converter.h"
-#include "util/enum-range.h"
-#include "util/finalizer.h"
 #include "util/int-char-converter.h"
 #include "view/display-messages.h"
 #include "wizard/spoiler-table.h"
 #include "wizard/tval-descriptions-table.h"
 #include "wizard/wizard-messages.h"
-#include "wizard/wizard-spells.h"
-#include "wizard/wizard-spoiler.h"
 #include "world/world.h"
 #include <algorithm>
 #include <fstream>
